requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
starlette==0.37.2
google-generativeai==0.8.3
